-checkbox styles and variables moved
