+++ packages/react/src/themes/teams/components/Checkbox/checkboxVariables.ts
@@ -1,8 +1,7 @@
-import { CheckboxVariables as BaseCheckboxVariables } from '../../../base/components/Checkbox/checkboxVariables'
 import { pxToRem } from '../../../../lib'
 import * as _ from 'lodash'
 
-export type CheckboxVariables = Partial<BaseCheckboxVariables> & {
+export type CheckboxVariables = {
   rootPadding: string
 
   textColor: string
@@ -10,13 +9,44 @@ export type CheckboxVariables = Partial<BaseCheckboxVariables> & {
 
   textColorHover: string
   borderColorHover: string
+  background: string
+  borderColor: string
+  borderStyle: string
+  borderRadius: string
+  borderWidth: string
+  checkboxColor: string
+  gap: string
+  margin: string
+  padding: string
+
+  toggleBackground: string
+  toggleBorderColor: string
+  toggleBorderStyle: string
+  toggleBorderRadius: string
+  toggleBorderWidth: string
+  toggleIndicatorColor: string
+  toggleMargin: string
+  togglePadding: string
+
+  checkedBackground: string
+  checkedBorderColor: string
+  checkboxCheckedColor: string
   checkedBackgroundHover: string
   checkedTextColor: string
   checkedIndicatorColor: string
-
-  toggleIndicatorSize: string
+  checkboxToggleCheckedColor: string
+  checkboxToggleCheckedBackground: string
+  checkboxToggleCheckedBorderColor: string
   toggleCheckedPadding: string
+  toggleIndicatorSize: string
 
+  disabledColor: string
+  disabledBackground: string
+  disabledBorderColor: string
+  disabledCheckboxColor: string
+  disabledToggleIndicatorColor: string
+  disabledToggleBackground: string
+  disabledToggleBorderColor: string
   disabledBackgroundChecked: string
   disabledCheckedIndicatorColor: string
 }
@@ -26,6 +56,24 @@ const padding = pxToRem(2)
 const defaultValue = 'red'
 
 export default (siteVars: any): CheckboxVariables => ({
+  checkboxColor: 'transparent',
+
+  toggleBackground: 'transparent',
+  toggleBorderColor: siteVars.colors.grey[300],
+  toggleBorderStyle: `solid`,
+  toggleBorderWidth: pxToRem(1),
+  toggleIndicatorColor: 'inherit',
+
+  checkboxCheckedColor: siteVars.colors.grey[500],
+  checkboxToggleCheckedBackground: 'transparent',
+  checkboxToggleCheckedBorderColor: siteVars.colors.grey[500],
+  checkboxToggleCheckedColor: 'inherit',
+
+  disabledColor: siteVars.colors.grey[300],
+
+  disabledCheckboxColor: siteVars.colors.grey[300],
+  disabledToggleBackground: 'transparent',
+  disabledToggleBorderColor: siteVars.colors.grey[200],
   textColor: _.get(siteVars, 'colorScheme.default.foreground1', defaultValue),
   background: 'transparent',
   borderColor: _.get(siteVars, 'colorScheme.default.foreground1', defaultValue),
